+++ devops/upgrade/0.0.6.1.sh
@@ -1,8 +1,8 @@
 #!/bin/bash
 
 # To be executed at the root of the project
+# To upgrade from 0.0.5.9 to 0.0.6.2
 
-# backup
 
 sudo apt update
 sudo apt -yy upgrade
@@ -15,11 +15,14 @@ git pull origin dev
 ./bin/install
 
 # Database upgrade
-#./bin/psql gargantext.ini < devops/postgres/upgrade/0.0.6.0.sql
-# exec script haskell upgrade
+echo "0.0.6.0 SQL upgrade"
+./bin/psql gargantext.ini < devops/postgres/upgrade/0.0.6.0.sql
 
-#./bin/psql gargantext.ini < devops/postgres/upgrade/0.0.6.1.sql
-# ~/.local/bin/gargantext-upgrade
+echo "Executing script haskell upgrade"
+~/.local/bin/gargantext-upgrade
+
+echo "0.0.6.1 SQL upgrade"
+./bin/psql gargantext.ini < devops/postgres/upgrade/0.0.6.1.sql
 
 sudo -i -u postgres bash << EOF 
 pg_dumpall > /tmp/backup.dump
@@ -29,23 +32,31 @@ sudo sed -i "s/bullseye/bookworm/g" /etc/apt/sources.list
 sudo apt update
 sudo apt -yy dist-upgrade
 sudo apt install -y postgresql-14 libpq-dev
+sudo apt remove --purge postgresql-11 postgresql-13
+sudo apt autoremove
+
 
 sudo -i -u postgres bash << EOF 
 psql < /tmp/backup.dump
 EOF
 
-sed -i "s/DB_PORT = 5432/DB_PORT = 5433/" gargantext.ini
+sed -i "s/DB_PORT = 5432/DB_PORT = 5434/" gargantext.ini
+
+# be sure the DB password is the right one
 DBPASS=$(grep "DB_PASS" gargantext.ini | sed "s/^.*= //")
+echo $DBPASS
 
-
-
-sudo -i -u postgres bash << EOF 
-psql < 'ALTER ROLE gargantua password \'$DBPASS\'';
+sudo -i -u postgres psql << EOF 
+ALTER ROLE gargantua password '${DBPASS}';
 EOF
 
-
+echo "0.0.6.2 SQL upgrade"
 ./bin/psql gargantext.ini < devops/postgres/upgrade/0.0.6.2.sql
 
+# Make sure compilation is ok
+./bin/install
+
+echo "Upgrade is over"