+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 * [BACK][FIX] Confluence on Graph
 * [BACK][FIX] HAL parser fix
 * [BACK][UPGRADE] ./bin/psql gargantext.ini < devops/postgres/upgrade/0.0.6.2.sql
-* [BACK][ADMIN] upgrade script (WIP)
+* [BACK][ADMIN] upgrade script, test ok
 
 ## Version 0.0.6.1
 * [BACK][FEAT] Spacy Server connection for French (and others) languages